requirements.txt
flask
flask-socketio
python-socketio
eventlet
ultralytics
opencv-python-headless
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
pillow
werkzeug
gunicorn
norfair